refactor: split agentWork repository modules
- Move work-item lifecycle into workItemStateRepository
- Move publish-record and check-run persistence into publishRecordRepository
- Keep repository.ts as a re-export barrel for existing import sites
